one process
This removes the parent-child process scheme that was used for process
monitoring. Registry-static isn't the the business of being a
process-monitoring tool, not should it be, so this commit removes all
that in favor of using well-used monitoring tools like monit, god,
upstart, systemd, etc.

As an added bonus, random stream errors are less likely to happen, and
the whole thing is easier to maintain.